Consolidate classes.
