+++ tests/data-access/app-database-context.test.ts
@@ -1,20 +1,17 @@
+import seedJson from "../data/json/database.json";
 import { ArmorItem } from "../../src/features/gear/armor-item";
 import { AppDatabaseContext } from "../../src/lib/data-access/app-database-context";
-import { SingleJsonFileDatabase } from "../../src/lib/data-access/single-json-file-database";
 import { AppSettings } from "../../src/lib/settings/app-settings";
-import { SeedableDatabaseStorage } from "../../src/lib/data-access/seedable-database-storage";
 import { BrowserBlobStorage } from "../../src/lib/data-access/browser-blob-storage";
-import seedJson from "../data/json/database.json";
+import { SeedableJsonDatabase } from "../../src/lib/data-access/seedable-json-database";
 
 describe("AppDatabaseContext", () => {
     let appSettings: AppSettings;
-    let databaseStorage: SeedableDatabaseStorage;
-    let database: SingleJsonFileDatabase;
+    let database: SeedableJsonDatabase;
 
     beforeEach(async () => {
         appSettings = new AppSettings();
-        databaseStorage = new SeedableDatabaseStorage(new BrowserBlobStorage(window.localStorage), seedJson);
-        database = new SingleJsonFileDatabase(databaseStorage);
+        database = new SeedableJsonDatabase(new BrowserBlobStorage(window.localStorage), seedJson);
 
         appSettings.connectionString = "app_db_key";
     });
